Asset manager updates - Add Asset<T> class for referencing assets - Support reloading assets. Used in rendered for reloading shaders.   - TODO: Maybe add a file watcher that reloads assets automatically? or at least prompts the user to reload? - Some new File/URI utilities (walk, isFile, isFolder, etc). - Update widgets   - Input for Colours   - Fix generic Input for vector types - ShaderPool is made obsolete by the AssetManager changes (really only provided reloading, which is now covered) - Fixed reference count bug for shaders and programs in GraphicsDevice_OpenGL - Fixed shadow mapping issue when the depth of the shaded pixel is past the far plane of the shadow frustum - project lua fixes

+++ orbital/lib/src/render/OpenGL/GraphicsDevice_OpenGL.cpp
@@ -295,7 +295,7 @@ namespace bfc {
 
     Vec3i TextureManager_OpenGL::getSize(GraphicsResource textureID, int64_t mipLevel) {
       Vec3i size = getTexture(textureID).size;
-      size /= (1ll << mipLevel);
+      size /= (1 << (int32_t)mipLevel);
       return {std::max(1, size.x), std::max(1, size.y), std::max(1, size.z)};
     }
